Changed Module Offsets
Changed module offsets to different values so that the giant names would
go away.  I feel that in future versions this feature will be more
refined, since in V7.7 was the very first version it appeared in.  The
manual inside of the Eagle install in C:/ thankfully is updated with the
new features and explains some of the new ways to do things better than
anywhere I can find online.  Seems that nobody online has these issues,
or I suppose that they just automatically think to read the 1000 page
document every update.