Fixed the board consistency by making a new board
Do not modify the schematic unless the schematic and board are
consistant.  The easiest way to do this is to just have the board open
when you open the schematic.  This is why it asks if you want to open
both when you first open the program.

Also added a fancy logo to the Atmel chip's footprint, also found an
existing part for the Xbee900 Pro, so I added that into the schematic
and made a module for it.

Also added in modules for the memory and GPS.  That will be what I work
on next.

Outstanding things:  I think the power module is a little funky with the
ports you have on there, Cooper. Are the two return lines not a common
ground?  Why separate them?  Also I recommend changing the ports on the
module to reflect any new net name changes, as well as having a
consistent direction of the pin, so set them to "pwr" when you create
them.  This will come in handy later with design rule checks, because
it'll flag any inconsistancies.