Added H3LIS100DL
Added the H3LIS100DL to the ST-microelectronics library along with a
3x3mm TFLGA-16 package. This device is in the HIGH_G module under
Sensors. Things for Evan to consider on this push:
1. Frame needed for the HIGH_G module schematic
2. Data communication for H3LIS100DL: SPI or I2C?
2a. If I2C, which address? Currently a resistor on the top level
schematic pulling it to 0.
3. Please check my footprint. Once again, I have more evidence for my
hate of ST: they did not provide a recommended solder footprint for the
TFLGA package. Therefore, I took the mechanical pad side of 0.35x0.25
and added 0.1 mm to each dimension to yield 0.45x0.35 (according to
their note TN0018) . I am fairly certain I centered the pads on the
pins. Yes, that is a purposeful solder mask expansion.