Ordnance & Breakwire modules fixes
1. In the Ordnance module: combined the PYRO+ and PWR to the same pin
and called PWR/PYRO+. This is the ordnance power from the super caps.
2. In the Ordnance module: combined CONT_STAT and PYRO– to the same pin
and called CONT_STAT/PYRO–. Outside the modules on the top level
schematic, I added long wires from this pin going to the left of each
module. This is also the digital line for sensing continuity status of
the ordnance lines.
3. In the Breakwire module, combined BW_STAT and BW– to the same pin and
called BW_STAT/BW–. Outside the module on the top level schematic, I
made long lines from this pin to the right of each module. This is the
digital input for sensing the breakwire status (open or closed).