Major revisions
1. Made board 30mm taller to allow for more space in cramped enviroment.
Deemed absolutely critical.
2. Majorly changed the power turn-on and turn-off MOSFETs design and
location to clear up room in the mid-board area
3. Moved the super capacitor nearer to the top of the extended section,
and also made it have a verticle orientation; should allow for better
handling of vertical Gs
4. Moved the super-capacitor charging circuitry in the extra space under
the super-cap.  Fits nicely in a small area now.
5. Created Vishay.lbr, Adafruit.lbr, nkk.lbr, and semtech.lbr
6. Added common SMT pads to Semtech.lbr and Vishay.lbr
7. Created MOSFETs in Vishay.lbr
8.Created MOSFETs in semtech.lbr
9. Completely routed the GPS and passives; just needs UART connections
to the uC now
10. Changed the DRC (Design Rules) to better reflect the capabilities of
Advanced Circuit's $33/board abilities.  Changes are mostly downsizing
the rules to 7mil of smallest hole, trace, and spacing between copper to
copper
11. Modified all vias in tight locations and low power consumption roles
to be a 7mil (0.18mm, rounded) drill hole